Neptune_Earth_porkchop.py
Changes were made to the c3 and time to better contour results by adjusting the departure, arrival and timespan for each the window can be adjusted. This process is manual but, it works. Using the auto contouring added with AI does allow for tighter graphs but, the hardcoded method works fine.